movidos el main page a un usercontrol, commit para tener una copia de seguridad en caso que algo salga mal
